MCR-2922 Add Neo4J make MCRNeo4JDatabaseDriver singleton and it only manages the connection with the neo4j Database extract static methods to own class
